Rename _health_ranking.py -> _row_factory.py
Module no longer does health ranking after compute_team_row was
moved to snapshot.py. Its three functions (raw_to_row, rank_rows,
band_from_health) construct and sort Row objects from RawMetrics.
